Move int-to-IDeriveSha logic under derivesha package
