+++ blockchain/init_derive_sha.go
@@ -17,23 +17,9 @@
 package blockchain
 
 import (
-	"github.com/klaytn/klaytn/blockchain/types"
 	"github.com/klaytn/klaytn/blockchain/types/derivesha"
 )
 
 func InitDeriveSha(deriveShaImpl int) {
-	switch deriveShaImpl {
-	case types.ImplDeriveShaOriginal:
-		logger.Info("Using DeriveShaOrig!")
-		types.InitDeriveSha(derivesha.DeriveShaOrig{})
-	case types.ImplDeriveShaSimple:
-		logger.Info("Using DeriveShaSimple!")
-		types.InitDeriveSha(derivesha.DeriveShaSimple{})
-	case types.ImplDeriveShaConcat:
-		logger.Info("Using DeriveShaConcat!")
-		types.InitDeriveSha(derivesha.DeriveShaConcat{})
-	default:
-		logger.Error("Undefined deriveShaImpl!! use DeriveShaOrig!")
-		types.InitDeriveSha(derivesha.DeriveShaOrig{})
-	}
+	derivesha.InitDeriveSha(deriveShaImpl)
 }
